Moved out InceptionTime tests
